Add custom error types
