+++ test/ServiceRegistryModules.Core.Tests/ServiceCollectionExtensions.Tests.cs
@@ -7,6 +7,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Hosting;
 using Moq;
+using ServiceRegistryModules.Exceptions;
 using ServiceRegistryModules.Internal;
 using Xunit;
 
@@ -141,7 +142,7 @@ public class ServiceCollectionExtensions_Should {
             => config.OfTypes(typeof(TestService1), typeof(Dependencies)));
 
         // Assert
-        action.Should().Throw<InvalidOperationException>()
+        action.Should().Throw<RegistryConfigurationException>()
             .Which.Message.Should().Be("The following registry types do not implement IRegistryModule: TestService1, Dependencies");
     }
 
@@ -208,7 +209,7 @@ public class ServiceCollectionExtensions_Should {
         var action = () => services.ApplyRegistries(config => config.UsingEnvironment(new TestService1()));
 
         // Assert
-        action.Should().Throw<InvalidOperationException>()
+        action.Should().Throw<RegistryConfigurationException>()
             .Which.Message.Should().Be("Environment object must implement IHostEnvironment");
     }
